Remove Electron
Electron was intended to be used for UIs and overlays, but I never got
around to that, so Electron was left around with no purpose. This commit
finally removes it.

I ALSO fixed a billion build-related errors. In the process, I added
"type": "module" to package.json, so now a certain version of Node is
required for top-level import/export statements.

The entirety of the Keyiano code itself should remain unchanged with
this commit (i.e. there should be no functional differences).